Make code actually work
I realized that it wasn't actually working. Ended up having to declare
SlackAPIUtils as a global node variable so that I could use it with the
`new Function()` syntax to create a function from a string template,
which I was able to use to create a callback of sorts that didn't rely
on the value at the time of the callback.